chore: replaced old components with new refactored versions
