+++ client/components/audioRecognition.js
@@ -4,96 +4,9 @@ import SpeechRecognition from 'react-speech-recognition'
 import { fetchPhrases, fetchDefinition, fetchTasks, removeTask, addTask } from '../store'
 import axios from 'axios'
 import GifPlayer from 'react-gif-player'
-import allEmotions from './emotionalComponents'
-
-let addedEmotion = {
-  happy: (<h1>Yay!</h1>),
-  sad: (<div>
-          <p>It's perfectly OK to have sad feelings at times. As long as they don't happen too often or last too long, sad feelings — like all emotions — are just a natural part of life.Everyone feels sad at times.</p>
-          <h4>Here are some positive ways to deal with sad feelings:</h4>
-          <ul>
-            <li><b>Notice how you feel and why:</b> Knowing your emotions helps you understand and accept yourself. If you feel sad, notice it — but don't dwell on it too long or give it too much drama. Show yourself a little understanding. Remind yourself that sadness will pass and you'll feel better.
-            </li>
-            <li><b>Bounce back from disappointments or failures:</b> When things don't go your way, don't give up! Stay in the game. There's always next time. Give yourself credit for trying.
-            </li>
-            <li><b>Get support:</b> Even the most capable kids need support. The people in your life who believe in you and care (like parents, friends, and teachers) can comfort you when you feel sad.
-            </li>
-          </ul>
-        </div>),
-  angry: (
-    <div>
-      <p>Anger is a normal and even healthy emotion — but it's important to deal with it in a positive way. Uncontrolled anger can hurt your health and your friendships.</p>
-      <h4>Here are some positive ways to deal with anger:</h4>
-      <ul>
-        <li><b>Think before you speak:</b> In the heat of the moment, it's easy to say something you'll later regret. Take a few moments to collect your thoughts before saying anything.
-        </li>
-        <li><b>Once you're calm, express your anger:</b> As soon as you're thinking clearly, tell the people you were angry at why you were angry. State your concerns and needs clearly, but try not to hurt their feelings.
-        </li>
-        <li><b>Identify possible solutions:</b> Work on fixing the issue that made you mad in the first place. Remind yourself that anger won't fix anything and might only make it worse.
-        </li>
-      </ul>
-    </div>
-  ),
-  nervous: (
-    <div>
-      <p>Being nervous is never fun or easy. You may feel your heart beating fast, your palms might sweat or feel clammy, and you may even feel a little bit shaky and out of control. All you need to do to calm yourself down is to remember that everyone gets nervous from time to time and that you’re ultimately in control of your mind and body. If you have the right attitude and have some tricks to help calm you, you’ll be able to get rid of those jitters in no time at all.</p>
-      <h4>Here are some positive ways to deal with nervousness:</h4>
-      <ul>
-        <li><b>Take a deep breath:</b> Sometimes, all you need to do to calm down a bit is to focus on the breath rising and falling from your body. Just stop what you’re doing and work on breathing in and breathing out deeply, and letting yourself take long, careful breaths instead of taking the shorter breaths people tend to take when they’re nervous.
-        </li>
-        <li><b>Distract yourself:</b> Though you can’t ignore your fears or worries forever, if you feel like there’s nothing you can do to address it except to worry more, then you may just want to take your mind off of it for a little while. Do something that you think will help you forget your worries and feel more at ease, such as, Reading or watching TV.
-        </li>
-        <li><b>Get the nervous energy out:</b> Try releasing the nervous energy you have through physical activity such as dancing, singing, or running.
-        </li>
-      </ul>
-    </div>
-  ),
-  lonely: (
-    <div>
-      <p>Don’t let yourself fall into the trap of believing that loneliness is forever. You might feel lonely today, this week, or even this month, but it doesn’t mean you are alone or that you have no one who cares for you. Like all feelings, loneliness is impermanent and it does not define who you are. Accept that you feel lonely, then focus on moving forward.</p>
-      <h4>Here are some positive ways to deal with lonliness:</h4>
-      <ul>
-        <li><b>Help Others:</b> Helping others eases loneliness because it makes us be less focused on ourselves. It could be a classmate or a relative who might benefit from some time hanging out with you.
-        </li>
-        <li><b>Do something creative:</b> Try a coloring book or a jigsaw puzzle. Or think outside the box and come up with something that is fun and soothing for you to do
-        </li>
-        <li><b>Sing:</b> It’s almost impossible to feel lonely when you’re singing. I’ve tried it, and it works. Sing solo or let your favorite singer keep you company as you sing together.
-        </li>
-      </ul>
-    </div>
-  )
-
-}
-
-function calculator(firstNumber, secondNumber, operation){
-
-  let answer
-  switch (operation){
-    case '+':
-    case 'plus':
-      answer = +(firstNumber) + +(secondNumber)
-      break
-    case '-':
-    case 'minus':
-      answer = firstNumber - secondNumber
-      break
-    case '*':
-    case 'times':
-    case 'multiplied':
-      answer = firstNumber * secondNumber
-      break
-    case '/':
-    case 'divided':
-      answer = firstNumber / secondNumber
-      break
-    default:
-      answer = 'sorry I did not understand you'
-  }
-
-  return answer
-
-}
-
+import EmotionalComponents from './emotionalComponents'
+import { checkForEmotion, checkForDefinition, checkForMath, calculate, checkForWeather, checkForList } from '../helperFunctions'
+import { MainResponse } from './responseComponents'
 
 class AudioRecognition extends Component {
   constructor(props) {
@@ -102,52 +15,42 @@ class AudioRecognition extends Component {
     this.getGeoLocation()
     this.props.loadPhraseData()
     this.props.loadToDoList()
-    this.feelings = ['happy', 'sad', 'tired', 'nervous', 'angry']
-    this.mathOperations = ['+', '-', '*', '/', 'plus', 'minus', 'times', 'multiplied', 'divided']
-    this.clickHandler = this.clickHandler.bind(this)
-    this.response = ''
-    this.videoUrl = ''
-    this.weather = ''
-    this.listening = false
-    this.typeOfResponse = ''
-    this.found = false
-    this.definitionHandler = this.definitionHandler.bind(this)
-    this.emotionHandler = this.emotionHandler.bind(this)
-    this.weatherHandler = this.weatherHandler.bind(this)
-    this.mathHandler = this.mathHandler.bind(this)
-    this.toDoListHandler = this.toDoListHandler.bind(this)
-    this.checkForEmotion = this.checkForEmotion.bind(this)
-    this.checkForDefinition = this.checkForDefinition.bind(this)
-    this.checkForMath = this.checkForMath.bind(this)
-    this.checkForWeather = this.checkForWeather.bind(this)
-    this.checkForList = this.checkForList.bind(this)
-    this.finishedAsync = false
-    this.addedMedia = ''
 
-
-    console.log('allemotions are', allEmotions)
-
-
-  }
-
-  componentWillReceiveProps(nextProps) {
-
-    if ((Object.keys(nextProps).length !== 0 && this.props.definition !== nextProps.definition) || this.props.toDoList.length !== nextProps.toDoList.length) this.finishedAsync = true
-
-  }
-
-  componentDidMount() {
-    if (this.props.dinosaur === 'stegosaurus') {
-      this.dinosaurGifUrl = 'https://drive.google.com/uc?export=download&id=1jwO0PLd1G4jNBQcbtsW3zDHsc1_K9Kf-'
-    } else if (this.props.dinosaur === 'tyrannosaurus') {
-      this.dinosaurGifUrl = 'https://drive.google.com/uc?export=download&id=10oYkrHB_q2plQJxzELy8EyKsheHEgEip'
-    } else {
-      this.dinosaurGifUrl = 'https://drive.google.com/uc?export=download&id=1G2eR26NW6DJGbUkAsSRsvafatiqzpKR1'
+    this.state = {
+      response: '',
+      videoUrl: '',
+      weather: '',
+      listening: false,
+      typeOfResponse: '',
+      found: false,
+      finishedAsync: false,
+      addedMedia: '',
+      dinosaurGifUrl: ''
     }
   }
 
+  isAsyncActionComplete = (nextProps) => (Object.keys(nextProps).length !== 0 && this.props.definition !== nextProps.definition) || this.props.toDoList.length !== nextProps.toDoList.length
 
-  async getGeoLocation() {
+  componentWillReceiveProps = (nextProps) => {
+    if (this.isAsyncActionComplete(nextProps)) this.setState({finishedAsync: true})
+  }
+
+  componentDidMount = () => {
+    let url
+    switch(this.props.dinosaur){
+      case 'stegosaurus':
+        url = 'https://drive.google.com/uc?export=download&id=1jwO0PLd1G4jNBQcbtsW3zDHsc1_K9Kf-'
+        break
+      case 'tyrannosaurus':
+        url = 'https://drive.google.com/uc?export=download&id=10oYkrHB_q2plQJxzELy8EyKsheHEgEip'
+        break
+      default: url = 'https://drive.google.com/uc?export=download&id=1G2eR26NW6DJGbUkAsSRsvafatiqzpKR1'
+    }
+    this.setState({dinosaurGifUrl: url})
+  }
+
+
+  getGeoLocation = async () => {
     let weatherUrl
     if (navigator.geolocation) {
 
@@ -155,297 +58,156 @@ class AudioRecognition extends Component {
         weatherUrl = `https://fcc-weather-api.glitch.me/api/current?lat=${position.coords.latitude}&lon=${position.coords.longitude}`
         axios.get(weatherUrl)
           .then((weatherData) => {
-            this.weather = weatherData
+            this.setState({weather: weatherData})
           })
       })
     }
   }
 
-  renderSwitch = (type) => {
-
-    switch (type) {
-      case 'feeling':
-        return (<div><h1>{this.response}</h1>
-          <iframe width="560" height="315" src={`${this.videoUrl}?autoplay=1`} allow="autoplay; encrypted-media" allowFullScreen /></div>)
-      case 'weather':
-        return (<div><h1>{this.response}</h1>
-          <img width="560" height="315" src={this.props.weatherImages[this.weather.data.weather[0].main]} /></div>)
-      case 'math':
-        return (
-          <h3>The answer is {this.response}</h3>
-        )
-      case 'list':
-        return (
-          <div>
-            <h3>The answer is {this.response}</h3>
-          </div>
-        )
-      case 'definition':
-        this.transcript = ''
-        return (
-          <div>
-            <h1>definition</h1>
-            {!this.finishedAsync
-              ? <p>Waiting...</p>
-              : (<div>
-                {window.speechSynthesis.speak(new SpeechSynthesisUtterance(this.props.definition.text))}
-                <p>{this.props.definition.text}</p>
-                <img height="150" src={this.props.definition.image} />
-              </div>)
-            }
-          </div>
-        )
-      default:
-        return (<h1>Hello Steven</h1>)
-    }
-  }
-
-
-  clickHandler() {
-    this.listening = !this.listening
-    this.found = false
-    this.finishedAsync = false
-    this.props.resetTranscript()
-    this.props.listening ? this.props.stopListening() : this.props.startListening()
-    this.addedMedia = ''
-    this.addedEmotion = ''
-  }
-
-  checkForEmotion(transcriptArr){
-
-    let spokenFeeling = this.feelings.find((feeling) => {
-      return transcriptArr.includes(feeling)
+  startTalking = (response, responseType) => {
+    this.props.stopListening()
+    window.speechSynthesis.speak(new SpeechSynthesisUtterance(response))
+    this.setState({
+      found: true,
+      typeOfResponse: responseType
     })
-    if (spokenFeeling) {
-      this.emotionHandler(spokenFeeling)
-
-    }
-
   }
 
-  checkForDefinition(transcriptArr){
-    let spokenDefinition = transcriptArr.find((currentWord) => {
-      return currentWord === 'define' || currentWord === 'definition'
-    })
-    if (spokenDefinition) {
-      let index
-      index = spokenDefinition === 'define'
-        ? transcriptArr.indexOf('define') + 1
-        : transcriptArr.indexOf('definition') + 2
-      this.definitionHandler(transcriptArr[index])
-    }
-
-  }
-
-  checkForMath(transcriptArr){
-    let spokenOperation = this.mathOperations.find((operation) => {
-      return transcriptArr.includes(operation)
-    })
-
-    if (spokenOperation) {
-      let index = transcriptArr.indexOf(spokenOperation)
-      let secondIndex = spokenOperation === 'divided' || spokenOperation === 'multiplied' ? index + 2 : index + 1
-      this.mathHandler(transcriptArr[index - 1], transcriptArr[secondIndex], spokenOperation)
-    }
-  }
-
-  checkForWeather(transcriptArr){
-    let spokenWeather = transcriptArr.find((currentWord) => {
-      return currentWord === 'weather' || currentWord === 'temperature'
-    })
-
-    if (spokenWeather) {
-      this.weatherHandler(this.weather)
-    }
-  }
-
-  checkForList(transcriptArr){
-    if (transcriptArr.includes('list')) {
-      let index = transcriptArr.indexOf('list')
-      if ((transcriptArr[index - 2] === 'to' && transcriptArr[index - 1] === 'do') || transcriptArr[index - 1] === 'to-do') {
-        this.toDoListHandler(transcriptArr, index)
-
-      }
-    }
-  }
-
-  emotionHandler(word) {
-
-    // can i add these to the redux store?
-    this.response = this.props.motivationalWords[word].response
-    this.videoUrl = this.props.motivationalWords[word].videoUrl
-    this.addedMedia = <iframe src={`${this.videoUrl}?autoplay=1`} allow="autoplay; encrypted-media" allowFullScreen />
-    window.speechSynthesis.speak(new SpeechSynthesisUtterance(this.response))
-    this.props.stopListening()
-    this.found = true
-    this.typeOfResponse = 'feeling'
-    this.typeOfEmotion = word
-
-  }
-
-  async definitionHandler(word) {
-    this.typeOfResponse = 'definition'
-    this.found = true
-    this.props.stopListening()
-
-    await this.props.loadDefinition(word)
-
-  }
-
-
-  weatherHandler(weather) {
-
-    if (weather) {
-
-      let fahrenheit = weather.data.main.temp * 1.8 + 32
-      fahrenheit = Math.round(fahrenheit).toString()
-      let percipitation = weather.data.weather[0].main === 'Clear' ? 'Clear Skies' : weather.data.weather[0].main
-      this.response = `It is ${fahrenheit} degrees fahrenheit outside with ${percipitation}`
-      this.addedMedia = <img height="150px" src={this.props.weatherImages[this.weather.data.weather[0].main]} />
-
-    } else {
-      this.response = 'The temperature and weather is currently unavailiable'
-    }
-
-    this.found = true
-    this.props.stopListening()
-    window.speechSynthesis.speak(new SpeechSynthesisUtterance(this.response))
-    this.typeOfResponse = 'weather'
-
-
-  }
-
-  mathHandler(firstNumber, secondNumber, operation) {
-
-    let answer = calculator(firstNumber, secondNumber, operation)
-
-    window.speechSynthesis.speak(new SpeechSynthesisUtterance(answer))
-    this.props.stopListening()
-    this.found = true
-    this.response = `The answer is ${answer}`
-    this.typeOfResponse = 'math'
-
-  }
-
-  toDoListHandler(arr, index) {
-    let modifierIndex
-    let endingIndex
-
-    if (arr.includes('add')) {
-      modifierIndex = arr.indexOf('add')
-      endingIndex = arr[index - 1] === 'to-do' ? index - 3 : index - 4
-
-      let newTask = arr.slice(modifierIndex + 1, endingIndex)
-
-      this.props.addToToDoList(newTask.join(' '))
-
-      this.response = `You  have just added ${newTask.join(' ')} to your to-do list`
-
-      window.speechSynthesis.speak(new SpeechSynthesisUtterance(this.response))
-      this.props.stopListening()
-      this.found = true
-      this.typeOfResponse = 'list'
-      this.listening = 'false'
-
-    } else if (arr.includes('remove') || arr.includes('delete')) {
-
-      modifierIndex = arr.indexOf('remove')
-      if (modifierIndex === -1) {
-        modifierIndex = arr.includes('delete')
-      }
-      endingIndex = arr[index - 1] === 'to-do' ? index - 3 : index - 4
-      let removedTask = arr.slice(modifierIndex + 1, endingIndex)
-
-
-      this.props.removeFromToDoList(removedTask.join(' ').toLowerCase())
-      this.response = `You  have just removed ${removedTask.join(' ')} from your to-do list`
-      window.speechSynthesis.speak(new SpeechSynthesisUtterance(this.response))
-
-
-    } else {
-
-      let list = this.props.toDoList.map((task) => task.task)
-      list = list.join(', ')
-
-      if (this.props.toDoList.length > 1) {
-        let lastIndex = list.lastIndexOf(',')
-        list = list.substring(0, lastIndex) + ' and ' + list.substring(lastIndex + 1)
-      }
-
-      list = `There are ${this.props.toDoList.length} things on your to do list. You should ${list}`
-      window.speechSynthesis.speak(new SpeechSynthesisUtterance(list))
-      this.response = list
-
-    }
-
-    this.props.stopListening()
-    this.found = true
-    this.typeOfResponse = 'list'
-    this.listening = 'false'
-
-  }
-
-
-  render() {
-
-    const { transcript, browserSupportsSpeechRecognition, listening } = this.props
+  parseCommand = (transcript, listening) => {
 
     let transcriptArr = transcript.split(' ')
 
     if (listening === true) {
       for (let word of transcriptArr) {
-
         if (word === 'please') {
 
-          transcriptArr = transcriptArr.map((currentWord) => {
-            return currentWord.toLowerCase()
-          })
-
-          this.checkForEmotion(transcriptArr)
-          this.checkForDefinition(transcriptArr)
-          this.checkForMath(transcriptArr)
-          this.checkForWeather(transcriptArr)
-          this.checkForList(transcriptArr)
-
+          transcriptArr = transcriptArr.map((currentWord) => currentWord.toLowerCase())
+          this.checkForKeyWords(transcriptArr)
         }
-
       }
+    }
+  }
 
+  checkForKeyWords = (transcriptArr) => {
+
+    let spokenEmotion = checkForEmotion(transcriptArr),
+      wordToDefine = checkForDefinition(transcriptArr),
+      mathOperationData = checkForMath(transcriptArr),
+      weather = checkForWeather(transcriptArr),
+      listData = checkForList(transcriptArr, this.props.toDoList)
+
+    if (spokenEmotion) this.emotionHandler(spokenEmotion)
+    else if (wordToDefine) this.definitionHandler(wordToDefine)
+    else if (mathOperationData) this.mathHandler(mathOperationData)
+    else if (weather) this.weatherHandler(this.state.weather)
+    else if (listData) this.toDoListHandler(listData)
+  }
+
+  clickHandler = () =>  {
+
+    this.props.resetTranscript()
+    this.props.listening ? this.props.stopListening() : this.props.startListening()
+  
+    this.setState({
+      addedEmotion: '',
+      addedMedia: '',
+      finishedAsync: false,
+      found: false,
+      listening: !this.state.listening
+    })
+  }
+
+  emotionHandler = (word) => {
+
+    this.setState({
+      response: this.props.motivationalWords[word].response,
+      videoUrl: this.props.motivationalWords[word].videoUrl,
+      addedMedia: <iframe src={`${this.props.motivationalWords[word].videoUrl}?autoplay=1`} allow="autoplay; encrypted-media" allowFullScreen />
+    })
+
+    this.startTalking(this.props.motivationalWords[word].response, 'feeling')
+    this.state.typeOfEmotion = word
+  }
+
+  definitionHandler = async (word) => {
+    //await this.props.loadDefinition(word)
+
+    // added due to api problems
+    this.props.definition.text = 'Sorry the dictionary feature is currently unavailable'
+    this.startTalking(this.props.definition.text, 'definition')
+
+    this.setState({ finishedAsync: true })
+  }
+
+
+  weatherHandler = (weather) => {
+    if (weather) {
+
+      let fahrenheit = weather.data.main.temp * 1.8 + 32
+      fahrenheit = Math.round(fahrenheit).toString()
+      let percipitation = weather.data.weather[0].main === 'Clear' ? 'Clear Skies' : weather.data.weather[0].main
+      this.state.response = `It is ${fahrenheit} degrees fahrenheit outside with ${percipitation}`
+      this.state.addedMedia = <img height="150px" src={this.props.weatherImages[this.state.weather.data.weather[0].main]} />
+
+    } else {
+      this.state.response = 'The temperature and weather is currently unavailiable'
+    }
+
+    this.startTalking(this.state.response, 'weather')
+  }
+
+  mathHandler = (operationObj) => {
+    let answer = calculate(operationObj)
+    this.state.response = `The answer is ${answer}`
+    this.startTalking(this.state.response, 'math')
+  }
+
+  toDoListHandler = taskInfo => {
+    let { response, action, task } = taskInfo
+
+    if (action === 'add'){
+      this.props.addToToDoList(task)
+    } else if (action === 'remove'){
+      this.props.removeFromToDoList(task)
+    }
+
+    this.state.response = response
+    this.startTalking(this.state.response, 'math')
+  }
+
+  render = () => {
+
+    const { transcript, browserSupportsSpeechRecognition, listening } = this.props
+    let weatherImage = this.state.weather.data ? this.props.weatherImages[this.state.weather.data.weather[0].main] : null
+    let responseData = {
+      type: this.state.typeOfResponse,
+      response: this.state.response,
+      finishedAsync: this.state.finishedAsync,
+      definition: this.props.definition.text,
+      addedMedia: this.state.addedMedia,
+      weatherImage: weatherImage,
+      emotionalResponse: EmotionalComponents[this.state.typeOfEmotion],
+      dictionaryImage: this.props.definition.image
     }
 
     if (!browserSupportsSpeechRecognition) {
       return null
     }
+    this.parseCommand(transcript, listening)
 
-      return (
-
-        <div id="audioPage">
-          <button id="audioPageButton" onClick={this.clickHandler}>{listening ? 'Stop' : 'Start'}</button>
-          <div id="audioUserBubble">
-            <h2>{transcript}</h2>
-          </div>
-          <div id="audioDinoBubble">
-            <div id="audioDinoPicture"><GifPlayer gif={this.dinosaurGifUrl} /></div>
-            {this.typeOfResponse !== 'definition'
-            ? (
-              <div>
-                <h2 id="audioDinoH2">{this.response}</h2>
-                <div className="responseImage">{this.addedMedia}</div>
-                <div>{this.typeOfResponse === 'feeling' ? addedEmotion[this.typeOfEmotion] : null}</div>
-              </div>
-            )
-            : (
-              <div>{this.renderSwitch(this.typeOfResponse)}</div>
-            )
-
-            }
-
-          </div>
+    return (
+      <div id="audioPage">
+        <button id="audioPageButton" onClick={this.clickHandler}>{listening ? 'Stop' : 'Start'}</button>
+        <div id="audioUserBubble">
+          <h2>{transcript}</h2>
         </div>
-
-      )
+        <div id="audioDinoBubble">
+          <div id="audioDinoPicture"><GifPlayer gif={this.state.dinosaurGifUrl} /></div>
+          {!this.state.typeOfResponse ? null : <MainResponse data={responseData} />}
+        </div>
+      </div>
+    )
   }
-
 }
 
 /**
@@ -465,7 +227,6 @@ const mapState = (state) => ({
       ThunderStorms: 'https://www.flyingmag.com/sites/flyingmag.com/files/styles/1000_1x_/public/images/2017/06/everything-explained-june.jpg?itok=UsjdV7uz&fc=50,50'
     }
 })
-
 
 const mapDispatch = dispatch => {
   return {
@@ -487,9 +248,4 @@ const mapDispatch = dispatch => {
   }
 }
 
-const options = {
-  autoStart: false
-}
-
-
-export default connect(mapState, mapDispatch)(SpeechRecognition(options)(AudioRecognition))
+export default connect(mapState, mapDispatch)(SpeechRecognition({autoStart: false})(AudioRecognition))
