+++ client/components/index.js
@@ -1,5 +1,3 @@
-export {default as NewAudioRecognition} from './newAudioRecognition'
-export {default as NewNewAudioRecognition} from './newNewAudioRecognition'
 export {default as AudioRecognition} from './AudioRecognition'
 export {default as Home} from './home'
 export {default as Choose} from './choose'
