+++ client/routes.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import {withRouter, Route, Switch} from 'react-router-dom'
-import {Home, About, Instructions, Choose, NewNewAudioRecognition, AudioRecognition } from './components'
+import {Home, About, Instructions, Choose, AudioRecognition } from './components'
 
 /**
  * COMPONENT
@@ -12,8 +12,7 @@ const Routes = () => (
     <Route exact path="/about" component={About} />
     <Route exact path="/instructions" component={Instructions} />
     <Route exact path="/choose" component={Choose} />
-    <Route exact path="/newAudio" component={NewNewAudioRecognition} />
-    <Route exact path="/oldAudio" component={AudioRecognition} />
+    <Route exact path="/audio" component={AudioRecognition} />
     <Route exact path="/" component={Home} />
   </Switch>
 )
